fix:清空话题总是修复当前话题
