+++ src/renderer/src/pages/home/Tabs/AssistantsTab.tsx
@@ -1,170 +1,23 @@
-import { DeleteOutlined, EditOutlined, MinusCircleOutlined, PlusOutlined, SaveOutlined } from '@ant-design/icons'
-import DragableList from '@renderer/components/DragableList'
-import CopyIcon from '@renderer/components/Icons/CopyIcon'
-import Scrollbar from '@renderer/components/Scrollbar'
-import { useAgents } from '@renderer/hooks/useAgents'
-import { useAssistant, useAssistants } from '@renderer/hooks/useAssistant'
-import { modelGenerating } from '@renderer/hooks/useRuntime'
-import { useSettings } from '@renderer/hooks/useSettings'
-import AssistantSettingsPopup from '@renderer/pages/settings/AssistantSettings'
-import { getDefaultTopic } from '@renderer/services/AssistantService'
-import { EVENT_NAMES, EventEmitter } from '@renderer/services/EventService'
-import { Assistant } from '@renderer/types'
-import { uuid } from '@renderer/utils'
-import { Dropdown } from 'antd'
-import { ItemType } from 'antd/es/menu/interface'
-import { last, omit } from 'lodash'
-import { FC, useCallback, useState } from 'react'
-import { useTranslation } from 'react-i18next'
-import styled from 'styled-components'
+import { useCallback, useState, FC } from "react"
+import { useTranslation } from "react-i18next"
+import { PlusOutlined } from '@ant-design/icons'
+import DragableList from "@renderer/components/DragableList"
+import Scrollbar from "@renderer/components/Scrollbar"
+import { useAgents } from "@renderer/hooks/useAgents"
+import { useAssistants } from "@renderer/hooks/useAssistant"
+import { Assistant } from "@renderer/types"
+import styled from "styled-components"
+import AssistantItemComponent from "@renderer/pages/home/Tabs/AssistantItemComponent"
 
-interface Props {
+// 类型定义
+interface AssistantsProps {
   activeAssistant: Assistant
   setActiveAssistant: (assistant: Assistant) => void
-  onCreateDefaultAssistant: () => void
   onCreateAssistant: () => void
+  onCreateDefaultAssistant: () => void
 }
 
-const Assistants: FC<Props> = ({
-  activeAssistant,
-  setActiveAssistant,
-  onCreateAssistant,
-  onCreateDefaultAssistant
-}) => {
-  const { assistants, removeAssistant, addAssistant, updateAssistants } = useAssistants()
-  const [dragging, setDragging] = useState(false)
-  const { removeAllTopics } = useAssistant(activeAssistant.id)
-  const { clickAssistantToShowTopic, topicPosition } = useSettings()
-  const { t } = useTranslation()
-  const { addAgent } = useAgents()
-
-  const onDelete = useCallback(
-    (assistant: Assistant) => {
-      const _assistant: Assistant | undefined = last(assistants.filter((a) => a.id !== assistant.id))
-      _assistant ? setActiveAssistant(_assistant) : onCreateDefaultAssistant()
-      removeAssistant(assistant.id)
-    },
-    [assistants, onCreateDefaultAssistant, removeAssistant, setActiveAssistant]
-  )
-
-  const getMenuItems = useCallback(
-    (assistant: Assistant) =>
-      [
-        {
-          label: t('assistants.edit.title'),
-          key: 'edit',
-          icon: <EditOutlined />,
-          onClick: () => AssistantSettingsPopup.show({ assistant })
-        },
-        {
-          label: t('assistants.copy.title'),
-          key: 'duplicate',
-          icon: <CopyIcon />,
-          onClick: async () => {
-            const _assistant: Assistant = { ...assistant, id: uuid(), topics: [getDefaultTopic(assistant.id)] }
-            addAssistant(_assistant)
-            setActiveAssistant(_assistant)
-          }
-        },
-        {
-          label: t('assistants.clear.title'),
-          key: 'clear',
-          icon: <MinusCircleOutlined />,
-          onClick: () => {
-            window.modal.confirm({
-              title: t('assistants.clear.title'),
-              content: t('assistants.clear.content'),
-              centered: true,
-              okButtonProps: { danger: true },
-              onOk: removeAllTopics
-            })
-          }
-        },
-        {
-          label: t('assistants.save.title'),
-          key: 'save-to-agent',
-          icon: <SaveOutlined />,
-          onClick: async () => {
-            const agent = omit(assistant, ['model', 'emoji'])
-            agent.id = uuid()
-            agent.type = 'agent'
-            addAgent(agent)
-            window.message.success({
-              content: t('assistants.save.success'),
-              key: 'save-to-agent'
-            })
-          }
-        },
-        { type: 'divider' },
-        {
-          label: t('common.delete'),
-          key: 'delete',
-          icon: <DeleteOutlined />,
-          danger: true,
-          onClick: () => {
-            window.modal.confirm({
-              title: t('assistants.delete.title'),
-              content: t('assistants.delete.content'),
-              centered: true,
-              okButtonProps: { danger: true },
-              onOk: () => onDelete(assistant)
-            })
-          }
-        }
-      ] as ItemType[],
-    [addAgent, addAssistant, onDelete, removeAllTopics, setActiveAssistant, t]
-  )
-
-  const onSwitchAssistant = useCallback(
-    async (assistant: Assistant) => {
-      await modelGenerating()
-
-      if (topicPosition === 'left' && clickAssistantToShowTopic) {
-        EventEmitter.emit(EVENT_NAMES.SWITCH_TOPIC_SIDEBAR)
-      }
-
-      setActiveAssistant(assistant)
-    },
-    [clickAssistantToShowTopic, setActiveAssistant, topicPosition]
-  )
-
-  return (
-    <Container className="assistants-tab">
-      <DragableList
-        list={assistants}
-        onUpdate={updateAssistants}
-        style={{ paddingBottom: dragging ? '34px' : 0 }}
-        onDragStart={() => setDragging(true)}
-        onDragEnd={() => setDragging(false)}>
-        {(assistant) => {
-          const isCurrent = assistant.id === activeAssistant?.id
-          return (
-            <Dropdown key={assistant.id} menu={{ items: getMenuItems(assistant) }} trigger={['contextMenu']}>
-              <AssistantItem onClick={() => onSwitchAssistant(assistant)} className={isCurrent ? 'active' : ''}>
-                <AssistantName className="name">{assistant.name || t('chat.default.name')}</AssistantName>
-                {isCurrent && (
-                  <MenuButton onClick={() => EventEmitter.emit(EVENT_NAMES.SWITCH_TOPIC_SIDEBAR)}>
-                    <TopicCount className="topics-count">{assistant.topics.length}</TopicCount>
-                  </MenuButton>
-                )}
-              </AssistantItem>
-            </Dropdown>
-          )
-        }}
-      </DragableList>
-      {!dragging && (
-        <AssistantItem onClick={onCreateAssistant}>
-          <AssistantName>
-            <PlusOutlined style={{ color: 'var(--color-text-2)', marginRight: 4 }} />
-            {t('chat.add.assistant.title')}
-          </AssistantName>
-        </AssistantItem>
-      )}
-      <div style={{ minHeight: 10 }}></div>
-    </Container>
-  )
-}
-
+// 样式组件（只定义一次）
 const Container = styled(Scrollbar)`
   display: flex;
   flex-direction: column;
@@ -184,18 +37,14 @@ const AssistantItem = styled.div`
   border-radius: var(--list-item-border-radius);
   border: 0.5px solid transparent;
   cursor: pointer;
-  .iconfont {
-    opacity: 0;
-    color: var(--color-text-3);
-  }
+
   &:hover {
     background-color: var(--color-background-soft);
   }
+
   &.active {
     background-color: var(--color-background-soft);
     border: 0.5px solid var(--color-border);
-    .name {
-    }
   }
 `
 
@@ -208,30 +57,60 @@ const AssistantName = styled.div`
   font-size: 13px;
 `
 
-const MenuButton = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  min-width: 22px;
-  height: 22px;
-  min-width: 22px;
-  min-height: 22px;
-  border-radius: 11px;
-  position: absolute;
-  background-color: var(--color-background);
-  right: 9px;
-  top: 6px;
-`
+const Assistants: FC<AssistantsProps> = ({
+  activeAssistant,
+  setActiveAssistant,
+  onCreateAssistant,
+  onCreateDefaultAssistant
+}) => {
+  const { assistants, removeAssistant, addAssistant, updateAssistants } = useAssistants()
+  const [dragging, setDragging] = useState(false)
+  const { addAgent } = useAgents()
+  const { t } = useTranslation()
 
-const TopicCount = styled.div`
-  color: var(--color-text);
-  font-size: 10px;
-  border-radius: 10px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-`
+  const onDelete = useCallback(
+    (assistant: Assistant) => {
+      const remaining = assistants.filter(a => a.id !== assistant.id)
+      const newActive = remaining[remaining.length - 1]
+      newActive ? setActiveAssistant(newActive) : onCreateDefaultAssistant()
+      removeAssistant(assistant.id)
+    },
+    [assistants, removeAssistant, setActiveAssistant, onCreateDefaultAssistant]
+  )
+
+  return (
+    <Container className="assistants-tab">
+      <DragableList
+        list={assistants}
+        onUpdate={updateAssistants}
+        style={{ paddingBottom: dragging ? '34px' : 0 }}
+        onDragStart={() => setDragging(true)}
+        onDragEnd={() => setDragging(false)}
+      >
+        {(assistant) => (
+          <AssistantItemComponent
+            key={assistant.id}
+            assistant={assistant}
+            isActive={assistant.id === activeAssistant.id}
+            onSwitch={setActiveAssistant}
+            onDelete={onDelete}
+            addAgent={addAgent}
+            addAssistant={addAssistant}
+            onCreateDefaultAssistant={onCreateDefaultAssistant}
+          />
+        )}
+      </DragableList>
+      {!dragging && (
+        <AssistantItem onClick={onCreateAssistant}>
+          <AssistantName>
+            <PlusOutlined style={{ color: 'var(--color-text-2)', marginRight: 4 }} />
+            {t('chat.add.assistant.title')}
+          </AssistantName>
+        </AssistantItem>
+      )}
+      <div style={{ minHeight: 10 }}></div>
+    </Container>
+  )
+}
 
 export default Assistants
